Playground 577

+++ src/PlaygroundGame/Service/Quiz.php
@@ -124,9 +124,33 @@ class Quiz extends Game implements ServiceManagerAwareInterface
             ErrorHandler::start();
             $image = $question->getImage();
             $image = str_replace($media_url, '', $image);
-            unlink($path .$image);
+            if (file_exists($path .$image)) {
+            	unlink($path .$image);
+            }
             $question->setImage(null);
             ErrorHandler::stop(true);
+        }
+        
+        $i = 0;
+        foreach ($question->getAnswers() as $answer) {
+        	if (!empty($data['answers'][$i]['upload_image']['tmp_name'])) {
+        		ErrorHandler::start();
+        		$data['answers'][$i]['upload_image']['name'] = $this->fileNewname($path, $question->getId() . "-" . $data['answers'][$i]['upload_image']['name']);
+        		move_uploaded_file($data['answers'][$i]['upload_image']['tmp_name'], $path . $data['answers'][$i]['upload_image']['name']);
+        		$answer->setImage($media_url . $data['answers'][$i]['upload_image']['name']);
+        		ErrorHandler::stop(true);
+        	}
+        	
+        	/*if(isset($data['delete_image']) && empty($data['upload_image']['tmp_name'])) {
+        		ErrorHandler::start();
+        		$image = $question->getImage();
+        		$image = str_replace($media_url, '', $image);
+        		unlink($path .$image);
+        		$question->setImage(null);
+        		ErrorHandler::stop(true);
+        	}*/
+        	
+        	$i++;
         }
 
         // Max points and correct answers recalculation for the quiz
@@ -188,7 +212,6 @@ class Quiz extends Game implements ServiceManagerAwareInterface
                         }
                     }
                 }
-
                 $winner = $this->isWinner($quiz, $quizCorrectAnswers);
                 $entry->setWinner($winner);
                 $entry->setPoints($quizPoints);
@@ -348,10 +371,9 @@ class Quiz extends Game implements ServiceManagerAwareInterface
                     $quizReply->addAnswer($quizReplyAnswer);
                     $quizPoints += 0;
                     $quizCorrectAnswers += 0;
-                    
                     $qAnswers = $question->getAnswers();
                     foreach ($qAnswers as $qAnswer) {
-                        if (trim(strip_tags($a)) == trim(strip_tags($qAnswer->getAnswer()))) {
+                       if (trim(strip_tags($a)) == trim(strip_tags($qAnswer->getAnswer()))) {
                             $quizReplyAnswer->setPoints($qAnswer->getPoints());
                             $quizPoints += $qAnswer->getPoints();
                             $quizReplyAnswer->setCorrect($qAnswer->getCorrect());
